Verification modals

+++ patients/views.py
@@ -1,11 +1,14 @@
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
+
+from homepage.forms import VerifyActionForm
 from patients.forms import NewPatient, SearchPatients, FlagForm
+from procedures.models import Procedure
 from roadmaps.forms import SelectFromRoadmap, RoadmapProcedureLinkForm
 from patients.models import Patients
 from assigned_procedures.models import AssignedProcedures
 from roadmaps.models import RoadmapProcedureLink
-import re
+from django.contrib import messages
 
 
 @login_required
@@ -31,11 +34,6 @@ def index(request):
                        'breadcrumbs': breadcrumbs})
 
 
-'''breadcrumbs = [('/roadmaps/', 'Roadmaps'),
-                          ('/roadmaps/view_roadmap/?id=' + roadmap_id, roadmap.roadmap_name),
-                          ('#', 'Modifying: ' + roadmap.roadmap_name)]'''
-
-
 @login_required
 def new_patient(request):
     if request.method == 'GET':
@@ -52,34 +50,13 @@ def new_patient(request):
             patient.save()
             return redirect('/homepage/')
         else:
-            breadcrumbs = [('/patients/', 'Patients'), ('#', 'New Patient')]
-            return render(request, 'new_patient.html',
-                          {'form': NewPatient(), 'failed_creation': True, 'title': 'New Patient',
-                           'breadcrumbs': breadcrumbs}, status=401)
+            messages.error(request, 'Invalid Form!')
+            return redirect('/patients/create/')
 
 
+# TODO This post currently takes you to the update patient page. Instead, this should be a seperate method that use request.GET
 @login_required
 def profile(request):
-    if request.method == 'POST':
-        try:
-            # Get desired patient id from url
-            patient = Patients.objects.get(id=request.GET.get('id'))
-            request.session['patient_id'] = patient.id
-
-            breadcrumbs = [('/patients/', 'Patients'),
-                           ('/patients/profile/?id=' + str(patient.id), patient.last_name + ', ' + patient.first_name),
-                           ('#', 'Update: ' + patient.last_name + ', ' + patient.first_name)]
-            return render(request, 'update_patient.html', {'form': NewPatient(
-                initial={'first_name': patient.first_name, 'last_name': patient.last_name,
-                         'record_number': patient.record_number,
-                         'birth_date': patient.bday, 'referring_physician': patient.referring_physician,
-                         'date_of_referral': patient.date_of_referral}), 'patient': patient,
-                'title': 'Update: ' + patient.last_name + ', ' + patient.first_name, 'breadcrumbs': breadcrumbs})
-
-        except Patients.DoesNotExist:
-            # TODO: add in error message here
-            return redirect('/patients/')
-
     if request.method == 'GET':
         try:
             # Get desired patient id from url
@@ -95,16 +72,39 @@ def profile(request):
                            'breadcrumbs': breadcrumbs, 'assigned_procedures': all_assigned_procedures,
                            'flag_form': FlagForm()})
         except Patients.DoesNotExist:
-            # TODO: add in error message here
+            messages.warning(request, "The patient you tried to reach doesn't exist!")
             return redirect('/patients/')
 
 
 @login_required
 def update(request):
+    if request.method == 'GET':
+        try:
+            # Get desired patient id from url
+            patient = Patients.objects.get(id=request.GET.get('id'))
+
+            breadcrumbs = [('/patients/', 'Patients'),
+                           ('/patients/profile/?id=' + str(patient.id), patient.last_name + ', ' + patient.first_name),
+                           ('#', 'Update: ' + patient.last_name + ', ' + patient.first_name)]
+
+            initial_form_dict = {'first_name': patient.first_name, 'last_name': patient.last_name,
+                                 'record_number': patient.record_number,
+                                 'birth_date': patient.bday, 'referring_physician': patient.referring_physician,
+                                 'date_of_referral': patient.date_of_referral}
+
+            page_title = 'Update: ' + patient.last_name + ', ' + patient.first_name
+
+            return render(request, 'update_patient.html',
+                          {'form': NewPatient(initial=initial_form_dict), 'patient': patient, 'title': page_title,
+                           'breadcrumbs': breadcrumbs, 'verification_form': VerifyActionForm()})
+
+        except Patients.DoesNotExist:
+            messages.warning(request, "The patient you tried to reach doesn't exist!")
+            return redirect('/patients/')
     if request.method == 'POST':
         form = NewPatient(request.POST)
         try:
-            patient = Patients.objects.get(id=request.session['patient_id'])
+            patient = Patients.objects.get(id=request.GET.get('id'))
             if form.is_valid():
                 # Clean form data and check that the username password pair is valid
                 cd = form.cleaned_data
@@ -118,29 +118,33 @@ def update(request):
                 patient.save()
                 return redirect("/patients/profile/?id=" + str(patient.id))
             else:
-                breadcrumbs = [('/patients/', 'Patients'),
-                               ('/patients/profile/?id=' + str(patient.id),
-                                patient.last_name + ', ' + patient.first_name),
-                               ('#', 'Update: ' + patient.last_name + ', ' + patient.first_name)]
-                return render(request, 'patient.html', {"patient": patient,
-                                                        'title': 'Profile: ' + patient.last_name + ', ' + patient.first_name,
-                                                        'failed_update': True, 'breadcrumbs': breadcrumbs}, status=401)
+                messages.error(request, 'Invalid Form!')
+                return redirect('/patients/profile/update/?id=' + str(patient.id))
         except Patients.DoesNotExist:
-            # TODO: add in error message here
+            messages.warning(request, "The patient you tried to reach doesn't exist!")
             return redirect('/patients/')
 
 
 @login_required
 def delete(request):
-    # TODO: This method can be dangerous if patient_id isnt properly set i think
     if request.method == 'POST':
+        form = VerifyActionForm(request.POST)
         try:
             # Get desired patient id from url
-            patient = Patients.objects.get(id=request.session['patient_id'])
-            patient.delete()
-            return redirect("/patients/")
+            patient = Patients.objects.get(id=request.GET.get('id'))
+            if form.is_valid():
+                cd = form.cleaned_data
+                if patient.record_number == cd['item_name']:
+                    patient.delete()
+                    return redirect('/patients/')
+                else:
+                    messages.error(request, 'Incorrect patient MRN!')
+                    return redirect('/patients/profile/update/?id=' + str(patient.id))
+            else:
+                messages.error(request, 'Invalid Form!2')
+                return redirect('/patients/profile/update/?id=' + str(patient.id))
         except Patients.DoesNotExist:
-            # TODO: add in error message here
+            messages.warning(request, "The patient you tried to reach doesn't exist!")
             return redirect('/patients/')
 
 
@@ -150,72 +154,91 @@ def procedures(request):
         try:
             # Get desired patient id from url
             patient = Patients.objects.get(id=request.GET.get('id'))
-            request.session['patient_id'] = patient.id
+
             roadmap_pairs = AssignedProcedures.get_all_procedures(patient)
+
             breadcrumbs = [('/patients/', 'Patients'),
                            ('/patients/profile/?id=' + str(patient.id), patient.last_name + ', ' + patient.first_name),
                            ('#', patient.first_name + " " + patient.last_name + "'s  Procedures")]
+
+            page_title = patient.first_name + " " + patient.last_name + "'s  Procedures"
+
             return render(request, 'patient_procedures.html',
-                          {'roadmap_form': SelectFromRoadmap(), 'procedure_phase_form': RoadmapProcedureLinkForm,
+                          {'roadmap_form': SelectFromRoadmap(),
+                           'procedure_phase_form': RoadmapProcedureLinkForm(),
+                           'patient': patient,
                            'breadcrumbs': breadcrumbs,
-                           'title': patient.first_name + " " + patient.last_name + "'s  Procedures",
+                           'title': page_title,
                            'roadmap_pairs': roadmap_pairs})
         except Patients.DoesNotExist:
-            # TODO: add in error message here
+            messages.warning(request, "The patient you tried to reach doesn't exist!")
             return redirect('/patients/')
 
 
+# TODO change request.session to getting the id from the url
 @login_required
 def add_roadmap(request):
     if request.method == 'POST':
         try:
-            patient = Patients.objects.get(id=request.session['patient_id'])
+            patient = Patients.objects.get(id=request.GET.get('id'))
 
             form = SelectFromRoadmap(request.POST)
             if form.is_valid():
                 cd = form.cleaned_data
                 roadmap = cd['roadmap']
                 AssignedProcedures.add_roadmap_to_patient(roadmap, patient)
-                return redirect('/patients/profile/?id=' + str(patient.id))
+                return redirect('/patients/profile/procedures/?id=' + str(patient.id))
             else:
+                messages.error(request, 'Invalid Form!')
                 return redirect('/homepage/')
         except Patients.DoesNotExist:
+            messages.warning(request, "The patient you tried to reach doesn't exist!")
             return redirect('/patients/')
 
 
+# TODO change request.session to getting the id from the url
 @login_required
 def add_procedure(request):
     if request.method == 'POST':
         try:
-            patient = Patients.objects.get(id=request.session['patient_id'])
+            patient = Patients.objects.get(id=request.GET.get('id'))
 
             form = RoadmapProcedureLinkForm(request.POST)
             if form.is_valid():
                 cd = form.cleaned_data
                 for procedure_item in cd['procedure']:
                     AssignedProcedures.assign_procedure_to_patient(cd['phase'], patient, procedure_item)
-                return redirect('/patients/profile/?id=' + str(patient.id))
+                return redirect('/patients/profile/procedures/?id=' + str(patient.id))
             else:
+                messages.error(request, 'Invalid Form!')
                 return redirect('/homepage/')
         except Patients.DoesNotExist:
+            messages.warning(request, "The patient you tried to reach doesn't exist!")
             return redirect('/patients/')
 
 
+# TODO change request.session to getting the id from the url
 @login_required
 def remove_pairs_from_patient(request):
     if request.method == 'POST':
         checked_boxes = request.POST.getlist('selection[]')
         try:
-            patient = Patients.objects.get(id=request.session['patient_id'])
-            for pair in checked_boxes:
-                cleaned_pair = tuple(pair.split(','))
-                AssignedProcedures.remove_assigned_procedure(patient, cleaned_pair[0],
-                                                             cleaned_pair[1])
-            return redirect('/patients/profile/?id=' + str(patient.id))
+            patient = Patients.objects.get(id=request.GET.get('id'))
+            if checked_boxes:
+                for pair in checked_boxes:
+                    cleaned_pair = tuple(pair.split(','))
+                    AssignedProcedures.remove_assigned_procedure(patient, cleaned_pair[0],
+                                                                 cleaned_pair[1])
+                return redirect('/patients/profile/procedures/?id=' + str(patient.id))
+            else:
+                #TODO Warning saying "You didn't select any items to be removed"
+                return redirect('/patients/profile/procedures/?id=' + str(patient.id))
         except Patients.DoesNotExist:
+            messages.warning(request, "The patient you tried to reach doesn't exist!")
             return redirect('/patients/')
 
 
+# TODO change request.session to getting the id from the url
 def flag_patient(request):
     if request.method == 'POST':
         try:
@@ -223,7 +246,6 @@ def flag_patient(request):
 
             if form.is_valid():
                 patient = Patients.objects.get(id=request.GET.get('id'))
-                request.session['patient_id'] = patient.id
 
                 cd = form.cleaned_data
 
@@ -235,9 +257,11 @@ def flag_patient(request):
                 return redirect('/patients/profile/?id=' + str(patient.id))
 
             else:
+                messages.error(request, 'Invalid Form!')
                 return redirect('/patients/')
 
         except Patients.DoesNotExist:
+            messages.warning(request, "The patient you tried to reach doesn't exist!")
             return redirect('/patients/')
 
 
@@ -245,7 +269,6 @@ def unflag_patient(request):
     if request.method == 'POST':
         try:
             patient = Patients.objects.get(id=request.GET.get('id'))
-            request.session['patient_id'] = patient.id
 
             patient.patient_flagged_reason = ""
             if patient.flagged:
@@ -255,4 +278,5 @@ def unflag_patient(request):
             return redirect('/patients/profile/?id=' + str(patient.id))
 
         except Patients.DoesNotExist:
+            messages.warning(request, "The patient you tried to reach doesn't exist!")
             return redirect('/patients/')
