+++ procedures/views.py
@@ -1,16 +1,10 @@
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
+
+from homepage.forms import VerifyActionForm
 from procedures.forms import NewProcedure, SearchProcedures
 from procedures.models import Procedure
-
-'''
-breadcrumbs = [('/patients/', 'Patients'),
-                               ('/patients/profile/?id=' + str(patient.id), patient.last_name + ', ' + patient.first_name),
-                               ('#', 'Update: ' + patient.last_name + ', ' + patient.first_name)]
-                               
-, 'breadcrumbs': breadcrumbs
-'''
-
+from django.contrib import messages
 
 @login_required
 def index(request):
@@ -37,6 +31,14 @@ def index(request):
 
 @login_required
 def new_procedure(request):
+    if request.method == 'GET':
+        breadcrumbs = [('/procedures/', 'Procedures'),
+                       ('#', 'New Procedure')]
+
+        return render(request, 'new_procedure.html',
+                      {'form': NewProcedure(initial={'time_frame': 'days'}), 'title': 'New Procedure',
+                       'breadcrumbs': breadcrumbs})
+
     if request.method == 'POST':
         form = NewProcedure(request.POST)
         if form.is_valid():
@@ -46,54 +48,55 @@ def new_procedure(request):
             procedure.save()
             return redirect('/procedures/')
         else:
-            breadcrumbs = [('/procedures/', 'Procedures'),
-                           ('#', 'New Procedure')]
-            return render(request, 'new_procedure.html',
-                          {'form': NewProcedure(initial={'time_frame': 'days'}), 'failed_creation': True, 'title': 'New Procedure',
-                           'breadcrumbs': breadcrumbs}, status=401)
-    else:
-        breadcrumbs = [('/procedures/', 'Procedures'),
-                       ('#', 'New Procedure')]
-        return render(request, 'new_procedure.html',
-                      {'form': NewProcedure(initial={'time_frame': 'days'}), 'title': 'New Procedure', 'breadcrumbs': breadcrumbs})
+            messages.error(request, 'Invalid Form!')
+            return redirect('/procedures/create/')
+
 
 
 @login_required
 def view_procedure(request):
-    if request.method == 'POST':
-        try:
-            # Get desired procedure id from url and store into session for when page is updated
-            procedure = Procedure.objects.get(id=request.GET.get('id'))
-            request.session['procedure_id'] = procedure.id
-            breadcrumbs = [('/procedures/', 'Procedures'),
-                           ('/procedures/view_procedure/?id=' + str(procedure.id),'View: ' + procedure.procedure_name),
-                           ('#', 'Update: ' + procedure.procedure_name)]
-            return render(request, 'update_procedure.html', {'form': NewProcedure(
-                initial={'procedure_name': procedure.procedure_name, 'notes': procedure.procedure_info, 'time_frame': 'days', 'time': procedure.est_days_to_complete}),
-                'procedure': procedure, 'title': 'Update: ' + procedure.procedure_name, 'breadcrumbs': breadcrumbs})
-        except Procedure.DoesNotExist:
-            return redirect('/procedures/')
-
     if request.method == 'GET':
         try:
             # Get desired patient id from url
             procedure = Procedure.objects.get(id=request.GET.get('id'))
             breadcrumbs = [('/procedures/', 'Procedures'),
                            ('#', 'View: ' + procedure.procedure_name)]
+
             return render(request, 'view_procedure.html',
                           {'procedure': procedure, 'title': 'View: ' + procedure.procedure_name,
                            'breadcrumbs': breadcrumbs})
         except Procedure.DoesNotExist:
-            # TODO: add in error message here
+            messages.warning(request, "The procedure you tried to reach doesn't exist!")
             return redirect('/procedures/')
 
 
 @login_required
 def update_procedure(request):
+    if request.method == 'GET':
+        try:
+            # Get desired procedure id from url and store into session for when page is updated
+            procedure = Procedure.objects.get(id=request.GET.get('id'))
+            
+            breadcrumbs = [('/procedures/', 'Procedures'),
+                           ('/procedures/view_procedure/?id=' + str(procedure.id), 'View: ' + procedure.procedure_name),
+                           ('#', 'Update: ' + procedure.procedure_name)]
+            
+            initial_form_data = {'procedure_name': procedure.procedure_name, 'notes': procedure.procedure_info,
+                         'time_frame': 'days', 'time': procedure.est_days_to_complete}
+            
+            return render(request, 'update_procedure.html', {'form': NewProcedure(initial= initial_form_data),
+                'procedure': procedure,
+                'title': 'Update: ' + procedure.procedure_name,
+                'breadcrumbs': breadcrumbs,
+                'verification_form': VerifyActionForm()})
+        except Procedure.DoesNotExist:
+            messages.warning(request, "The procedure you tried to reach doesn't exist!")
+            return redirect('/procedures/')
+
     if request.method == 'POST':
         form = NewProcedure(request.POST)
         try:
-            procedure = Procedure.objects.get(id=request.session['procedure_id'])
+            procedure = Procedure.objects.get(id=request.GET.get('id'))
             if form.is_valid():
                 # Clean form data and check that the username password pair is valid
                 cd = form.cleaned_data
@@ -103,26 +106,34 @@ def update_procedure(request):
                 procedure.add_time_estimate(cd['time'], str(request.POST.get('time_frame')))
                 procedure.save()
                 return redirect("/procedures/view_procedure/?id=" + str(procedure.id))
+
             else:
-                breadcrumbs = [('/procedures/', 'Procedures'),
-                               ('/procedures/view_procedure/?id=' + str(procedure.id),'View: ' + procedure.procedure_name),
-                               ('#', 'Update: ' + procedure.procedure_name)]
-                return render(request, 'view_procedure.html',
-                              {'procedure': procedure, 'title': 'View: ' + procedure.procedure_name,
-                               'failed_update': True, 'breadcrumbs': breadcrumbs}, status=401)
+                messages.error(request, 'Invalid Form!')
+                return redirect('/procedures/view_procedure/update/?id=' + str(procedure.id))
         except Procedure.DoesNotExist:
-            # TODO: add in error message here
+            messages.warning(request, "The procedure you tried to reach doesn't exist!")
             return redirect('/procedures/')
 
 
 @login_required
 def delete_this_procedure(request):
     if request.method == 'POST':
+        form = VerifyActionForm(request.POST)
         try:
+            procedure = Procedure.objects.get(id=request.GET.get('id'))
             # Get desired patient id from url
-            procedure = Procedure.objects.get(id=request.session['procedure_id'])
-            procedure.delete()
-            return redirect("/procedures/")
+            if form.is_valid():
+                cd = form.cleaned_data
+                if procedure.procedure_name == cd['item_name']:
+                    procedure.delete()
+                    return redirect('/procedures/')
+                else:
+                    messages.error(request, 'Incorrect procedure name!')
+                    return redirect('procedures/view_procedure/update/?id=' + str(procedure.id))
+            else:
+                messages.error(request, 'Invalid Form!')
+                return redirect('procedures/view_procedure/update/?id=' + str(procedure.id))
+
         except Procedure.DoesNotExist:
-            # TODO: add in error message here
+            messages.warning(request, "The procedure you tried to reach doesn't exist!")
             return redirect('/procedures/')
